+++ CHANGELOG.md
@@ -5,7 +5,7 @@ Unreleased
 ----------
 
 ### Added
-* Added an environment variable to support WebdriverIO's bail option.
+* Added an environment variable `WDIO_BAIL` to support WebdriverIO's bail option.
 
 ### Changed
 * Updated devtool sourceMap from cheap-source-map to eval-source-map